Refactor window tracking and add Git security improvements
- Add path sanitization to GitRepositoryMonitor for shell safety
  - Validate paths are absolute and exist before use
  - Escape single quotes to prevent shell injection
  - Add proper error handling for Git operations

- Fix SessionDetailView environment injection
  - Add missing gitRepositoryMonitor environment

- Refactor WindowTracker into modular components
  - Extract PermissionChecker for accessibility checks
  - Create ProcessTracker for process management
  - Add WindowEnumerator for window discovery
  - Implement WindowFocuser for window activation
  - Add WindowMatcher for window identification

- Refactor VibeTunnelMenuView into smaller components
  - Extract SessionRow, GitRepositoryRow, ServerInfoSection
  - Create MenuActionBar, MenuStyles, EmptySessionsView
  - Improve code organization and maintainability

- Add flexible Git path detection
  - Check multiple common Git binary locations
  - Support Homebrew and system installations
